changes in shipper for getall and openAd api

+++ DPUWebProject/src/main/java/com/dpu/model/CompanyResponse.java
@@ -14,7 +14,7 @@ public class CompanyResponse implements Serializable{
 	private static final long serialVersionUID = 1L;
 
 	//@JsonProperty("company_id")
-	private Integer companyId;
+	private Long companyId;
 	
 	//@JsonProperty("name")
 	private String name;
@@ -85,11 +85,12 @@ public class CompanyResponse implements Serializable{
 		this.billingLocations = billingLocations;
 	}
 
-	public Integer getCompanyId() {
+
+	public Long getCompanyId() {
 		return companyId;
 	}
 
-	public void setCompanyId(Integer companyId) {
+	public void setCompanyId(Long companyId) {
 		this.companyId = companyId;
 	}
